Safari-611.1.10/LayoutTests/fast/picture/viewport-resize-expected.txt

Test that pictures properly update when the viewport changes size.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
PASS innerWidth is 800
PASS innerHeight is 600
PASS document.getElementById("simple").clientWidth is 800
PASS currentSrcFileName("simple") is "image-set-2x.png?2"
PASS document.getElementById("multiple").clientWidth is 800
PASS currentSrcFileName("multiple") is "image-set-2x.png?2"
PASS document.getElementById("other_children").clientWidth is 800
PASS currentSrcFileName("other_children") is "image-set-2x.png?2"
PASS document.getElementById("img_before_source").clientWidth is 75
PASS currentSrcFileName("img_before_source") is "blue_rect.jpg?1"
PASS document.getElementById("unknown_type").clientWidth is 1600
PASS currentSrcFileName("unknown_type") is "image-set-4x.png?3"
PASS document.getElementById("svg_type").clientWidth is 800
PASS currentSrcFileName("svg_type") is "image-set-2x.png?2"
PASS document.getElementById("empty_type").clientWidth is 800
PASS currentSrcFileName("empty_type") is "image-set-2x.png?2"
PASS document.getElementById("whitespace_type").clientWidth is 800
PASS currentSrcFileName("whitespace_type") is "image-set-2x.png?2"
PASS document.getElementById("parameter_type").clientWidth is 800
PASS currentSrcFileName("parameter_type") is "image-set-2x.png?2"
PASS document.getElementById("empty_srcset").clientWidth is 1600
PASS currentSrcFileName("empty_srcset") is "image-set-4x.png?3"
PASS document.getElementById("no_srcset").clientWidth is 1600
PASS currentSrcFileName("no_srcset") is "image-set-4x.png?3"
PASS innerWidth is 1200
PASS innerHeight is 600
PASS document.getElementById("simple").clientWidth is 1600
PASS currentSrcFileName("simple") is "image-set-4x.png?3"
PASS document.getElementById("multiple").clientWidth is 1600
PASS currentSrcFileName("multiple") is "image-set-4x.png?3"
PASS document.getElementById("other_children").clientWidth is 1600
PASS currentSrcFileName("other_children") is "image-set-4x.png?3"
PASS document.getElementById("img_before_source").clientWidth is 75
PASS currentSrcFileName("img_before_source") is "blue_rect.jpg?1"
PASS document.getElementById("unknown_type").clientWidth is 1600
PASS currentSrcFileName("unknown_type") is "image-set-4x.png?3"
PASS document.getElementById("svg_type").clientWidth is 1600
PASS currentSrcFileName("svg_type") is "image-set-4x.png?3"
PASS document.getElementById("empty_type").clientWidth is 1600
PASS currentSrcFileName("empty_type") is "image-set-4x.png?3"
PASS document.getElementById("whitespace_type").clientWidth is 1600
PASS currentSrcFileName("whitespace_type") is "image-set-4x.png?3"
PASS document.getElementById("parameter_type").clientWidth is 1600
PASS currentSrcFileName("parameter_type") is "image-set-4x.png?3"
PASS document.getElementById("empty_srcset").clientWidth is 1600
PASS currentSrcFileName("empty_srcset") is "image-set-4x.png?3"
PASS document.getElementById("no_srcset").clientWidth is 1600
PASS currentSrcFileName("no_srcset") is "image-set-4x.png?3"

some text
